Get ontology from GitHub and import it via import_ontology.

Get ontology from GitHub and import it via import_ontology.

get_ontology_github(
  name,
  repo,
  filetype = ".obo",
  file = paste0(name, filetype),
  tag = "latest",
  save_dir = cache_dir(),
  force_new = FALSE,
  ...
)

Arguments

name

<...>Any ontology name from get_ols_options

"mondo" Import the Mondo ontology. GitHub..

"hpo" Import the Human Phenotype Ontology. GitHub.

repo

Repository name in format "owner/repo". Defaults to guess_repo().

filetype

File type to search for.

file

Can be one of the following:

"ontology"Creates an ontology_DAG R object by importing the OBO file directly from the official uPheno GitHub repository.

"bestmatches"Returns a merged table with the best matches between human and non-human homologous phenotypes (from multiple species). Distributed by the official uPheno GitHub repository.

"upheno_mapping"Return a merged table with matches between human and non-human homologous phenotypes (from multiple species). Distributed by the Monarch Initiative server.

tag

tag for the GitHub release to which this data should be attached.

save_dir

Directory to save a file to.

force_new

If TRUE, force a new download.

...

Arguments passed on to simona::import_ontology

verbose

Whether to print messages.

robot_jar

The path of the robot.jar file. It can be downloaded from https://github.com/ontodev/robot/releases. Internally, the file is converted to the obo format and parsed by import_obo(). The value of robot_jar can be set as a global option simona_opt$robot_jar = ....

JAVA_ARGS

Options for java. For example you can set -Xmx20G if you want to increase the memory to 20G for java.

Value

ontology_DAG