Map uPheno data

Get uPheno cross-species mapping data by:

Downloading cross-species phenotype-phenotype mappings.

Downloading within-spceies phenotype-gene mappings, and converting these genes to human orthologs.

Merging the phenotype-phenotype and phenotype-gene mappings.

Filtering out any mappings that do not have a human ortholog within each respective phenotype.

Calculating the proportion of orthologous genes overlapping across the species-specific phenotype-gene mappings.

Iterating the above steps using multiple methods (pheno_map_method) and concatenating the results together.

map_upheno_data(
  pheno_map_method = c("upheno", "monarch"),
  gene_map_method = c("monarch"),
  keep_nogenes = FALSE,
  fill_scores = NULL,
  terms = NULL,
  save_dir = cache_dir(),
  agg = TRUE,
  force_new = FALSE
)

Arguments

pheno_map_method

Method to use for mapping phenotypes across ontologies.

"upheno"Use uPheno's phenotype-to-phenotype mappings. Contains fewer ontologies but with greater coverage of phenotypes.

"monarch"Use Monarch's phenotype-to-phenotype mappings. Contains more ontologies but with less coverage of phenotypes.

gene_map_method

Method to use for mapping genes across species.

"monarch"Use Monarch's gene-to-gene mappings.

keep_nogenes

Logical indicating whether to keep mappings that do not have any orthologous genes.

fill_scores

Fill missing scores in the "equivalence_score" and "subclass_score" columns with this value. These columns represent the quality of mapping between two phenotypes on a scale from 0-1.

terms

A subset of HPO IDs to include in the final dataset and plots (e.g. c("HP:0001508","HP:0001507")).

save_dir

Directory to save cached data.

agg

Aggregate the data to to phenotype-level (TRUE). Otherwise, keep data at gene level (FALSE).

force_new

Create a new file instead of using any cached files.

Value

A data.table containing the mapped data.

Examples

if (FALSE) { # \dontrun{
pheno_map_genes_match <- map_upheno_data()
} # }